backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
neo4j==5.28.0
pydantic==2.8.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pandas==2.3.2
openpyxl==3.1.3
ipykernel==6.30.1
graphdatascience==1.16